chore: Modified dependencies, added comments and implemented compute_macs function.
Most of the imports across the algorithm folder have been modified, as
most of the imports were relative imports and lead to errors.

Other than that, the code as been commented to explain the
implementation.

I have introduced a function for computing the macs required for
performing a backward pass.